refactor(#9): extend dark mode with hook
