+++ util/useDarkMode.tsx
@@ -1,8 +1,9 @@
 'use client'
 
-import { useEffect, useMemo } from 'react'
+import { useEffect, useMemo, useState } from 'react'
 
-export const DarkMode = () => {
+export function useDarkMode() {
+  const [isDark, setIsDark] = useState(false)
   const mql = useMemo(() => {
     if (typeof matchMedia === 'undefined') {
       return
@@ -16,8 +17,10 @@ export const DarkMode = () => {
     }
 
     // toggle class attribute bsed on media query
-    const onChangeHandler = (e: MediaQueryList | MediaQueryListEvent) =>
+    const onChangeHandler = (e: MediaQueryList | MediaQueryListEvent) => {
+      setIsDark(e.matches)
       document.querySelector('html')?.classList.toggle('dark', e.matches)
+    }
     onChangeHandler(mql)
 
     // listen for changes
@@ -27,5 +30,10 @@ export const DarkMode = () => {
     return () => mql.removeEventListener('change', onChangeHandler)
   }, [mql])
 
+  return isDark
+}
+
+export const DarkMode = () => {
+  useDarkMode()
   return null
 }
